requirements.txt
pdfplumber
python-docx
spacy
scikit-learn
sentence-transformers
numpy
nltk
fastapi
uvicorn
python-multipart
streamlit